refactor : 검색 api 사용시 정규식 패턴 적용 로직 메서드 분리
