+++ routes/location.js
@@ -4,9 +4,9 @@ const {checkAuthenticated} = require("../middlewares/authMiddleware");
 const upload = require("../middlewares/s3Middleware");
 const router = express.Router();
 
-router.get("/",locationController.readLocations)
-router.post("/",checkAuthenticated,upload.single("image"),locationController.applyLocation)
-router.put("/:id",checkAuthenticated,upload.single("image"),locationController.updateLocation)
-router.get("/:id",locationController.readLocation);
+router.get("/", locationController.readLocations);
+router.get("/:id", locationController.readLocation);
+router.post("/", checkAuthenticated, upload.single("image"), locationController.applyLocation);
+router.put("/:id", checkAuthenticated, upload.single("image"), locationController.updateLocation);
 
 module.exports = router;
