+++ utils/LocationUtil.js
@@ -1,0 +1,5 @@
+exports.buildSearchRegex = (searchParam) => {
+    const searchTermWithoutSpaces = searchParam.replace(/\s/g, '');
+    const searchRegexString = searchTermWithoutSpaces.split('').join('.*');
+    return new RegExp(searchRegexString, 'i');
+}
